requirements.txt
requests==2.25.1
uvicorn==0.13.4
fastapi==0.63.0
opentelemetry-api~=1.8.0
opentelemetry-sdk~=1.8.0
opentelemetry-distro~=0.27b0
opentelemetry-exporter-otlp~=1.8.0
opentelemetry-instrumentation~=0.27b0
prometheus_fastapi_instrumentator~=5.7.1
opentelemetry-instrumentation-fastapi~=0.27b0